+++ src/components/ShoppingBag/ShoppingBagDishCard/ShoppingBagDishCard.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/mediaQueries.scss";
+@import "../../../styles/mediaQueries.scss";
 
 .ShoppingBagDishCardLayout {
   display: flex;
